// client_perfiles_huespedes/src/app/models/huesped.ts
export interface Huesped {
    id?: number;
    tipo?: string;
    nombres?: string;
    apellidos?: string;
    cedula?: string;
    telefono?: number;
    pais?: string;
    ciudad?: string;
    direccion?: string;
    enabled?: boolean;
    email?: string;
    username?: string;
    password?: string;
    createAt?: Date;
  }